+++ server/server.js
@@ -5,7 +5,6 @@ dotenv.config();
 import express from 'express';
 const app = express();
 const port = 3000;
-import { buySymbols } from './src/trade/buy/buy.js';
 
 app.get('/', (req, res) => {
     res.send('Hello world!');
@@ -14,5 +13,4 @@ app.get('/', (req, res) => {
 app.listen(port, () => {
     console.log(`Web app listening at http://localhost:${port}`);
     console.log(`env: ${process.env.ENV}`);
-    buySymbols();
 });
